backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
supabase
gspread
pytesseract
openai-whisper
python-multipart
sqlalchemy
python-dotenv
